pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "edgar-scraper"
version = "0.1.0"
description = "SEC EDGAR DEF 14A filing scraper"
requires-python = ">=3.11"
authors = [
    {name = "Shane"}
]
dependencies = [
    "requests",
    "pandas",
    "beautifulsoup4",
    "lxml",
    "backoff",
    "tqdm",
    "python-dotenv",
    "openai",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.hatch.build.targets.wheel]
packages = ["edgar_scraper.py", "parse_exec_compensation.py", "schema.py"]
